smith_1/requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
duckduckgo-search
python-telegram-bot==20.7
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
anthropic==0.34.2
cachetools==5.5.0
base58==2.1.1
pymongo[srv]==3.11
jinja2==3.1.3
python-multipart==0.0.6
pydantic==1.10.7
aiofiles==0.8.0
pycoingecko==3.1.0
web3==6.13.0
tweepy==4.14.0